swap connector naming with plugin
