+++ Makefile
@@ -1,4 +1,6 @@
 .PHONY: build
+test:
+	go test ./...
 build:
 	go vet ./...
 	go test ./...
