+++ internal/api/task_queues.go
@@ -3,5 +3,5 @@ package api
 const (
 	TestQueueFormat = "test.%s.%d"
 
-	CommonConnectorsQueue = "common.connectors"
+	CommonPluginsQueue = "common.plugins"
 )
